refactor: Mamba2Kernel + GDNKernel own perf data; Mamba2 composite moved (AIC-539)
Phase 3.3 of the SDK refactor. Three classes move out of ``_legacy.py``
into ``operations/mamba.py``:
- ``Mamba2Kernel`` (single Mamba2 kernel: conv1d / SSM) — owns
  ``_data_cache`` for ``mamba2_perf.txt``. ``PerfDatabase.query_mamba2``
  delegates here.
- ``GDNKernel`` (Gated DeltaNet kernel for Qwen3.5 linear attention) —
  owns ``_data_cache`` for ``gdn_perf.txt``. ``PerfDatabase.query_gdn``
  delegates here.
- ``Mamba2`` (composite for NemotronH hybrid models) — no CSV data of
  its own; calls ``database.query_gemm`` for in_proj + out_proj and
  ``database.query_mem_op`` for conv1d / SSM / norm. The
  ``query_mem_op`` callers are deliberately preserved — ``query_mem_op``
  is an explicit non-goal of this refactor (see handoff Decision #7).

What changes in ``perf_database.py``:
- ``__init__`` no longer eagerly loads ``_mamba2_data`` / ``_gdn_data``;
  ``Mamba2Kernel.load_data(self)`` and ``GDNKernel.load_data(self)`` are
  called right after the comm + DSA loads.
- ``query_mamba2`` / ``query_gdn`` reduced to one-line delegations.
- No extrapolation blocks or SOL clamping to remove — the legacy
  ``__init__`` / ``_correct_data`` paths never touched these tables
  (they're keyed by structural config tuples rather than dense
  ``(num_heads, s, b)`` grids).

Behavior preservation:
- Both table query paths preserve the full SOL fallback chain: missing
  data → SOL; missing kernel_source → SOL; missing phase → SOL;
  missing model_key with no d_model match → SOL; etc. ``"sol"`` /
  ``"silicon"`` / ``"empirical"`` source tags emitted as before.
- The nearest-config fallback for Mamba2 (first match by d_model) and
  GDN (closest match by abs(num_v_heads delta)) is preserved verbatim.
- The chunked-scan ``h_chunks_bytes`` accounting in GDN's
  ``chunk_gated_delta_rule`` SOL path is byte-for-byte unchanged.
- ``Mamba2.query`` composite arithmetic + ``"mixed"`` source-tag
  bookkeeping (when sub-results disagree on source) preserved.

Cache key matches every other migrated op:
``(systems_root, system, backend, version, enable_shared_layer)``.

Verification:
- 670 unit tests pass + 1 skipped.
- 38 regression snapshots bit-identical + 6 skipped (same as Phase 3.4 baseline).
- ruff check + format clean.
