+++ src/aiconfigurator/sdk/operations/_legacy.py
@@ -912,293 +912,6 @@ class WideEPContextMLA(Operation):
         return self._weights * self._scale_factor
 
 
-class Mamba2Kernel(Operation):
-    """
-    Single Mamba2 kernel op (Conv1D or SSM) using collected mamba2_perf data.
-
-    One of four kernels: causal_conv1d_fn, mamba_chunk_scan_combined (context),
-    causal_conv1d_update, selective_state_update (generation).
-    Uses full (unsharded) dimensions for lookup; collector data is per-layer.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        kernel_source: str,
-        phase: str,
-        hidden_size: int,
-        nheads: int,
-        head_dim: int,
-        d_state: int,
-        d_conv: int,
-        n_groups: int,
-        chunk_size: int,
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._kernel_source = kernel_source
-        self._phase = phase
-        self._hidden_size = hidden_size
-        self._nheads = nheads
-        self._head_dim = head_dim
-        self._d_state = d_state
-        self._d_conv = d_conv
-        self._n_groups = n_groups
-        self._chunk_size = chunk_size
-        self._weights = 0.0
-
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        batch_size = kwargs.get("batch_size")
-        s = kwargs.get("s")
-        seq_len = s if self._phase == "context" else None
-        result = database.query_mamba2(
-            phase=self._phase,
-            kernel_source=self._kernel_source,
-            batch_size=batch_size,
-            seq_len=seq_len,
-            d_model=self._hidden_size,
-            d_state=self._d_state,
-            d_conv=self._d_conv,
-            nheads=self._nheads,
-            head_dim=self._head_dim,
-            n_groups=self._n_groups,
-            chunk_size=self._chunk_size,
-        )
-        return PerformanceResult(
-            latency=float(result) * self._scale_factor,
-            energy=result.energy * self._scale_factor,
-            source=getattr(result, "source", "silicon"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-
-class GDNKernel(Operation):
-    """
-    Single Gated DeltaNet (GDN) kernel op for Qwen3.5 linear_attention layers.
-
-    Covers four kernel sources:
-      Context phase:
-        - "causal_conv1d_fn": Causal 1D convolution over full sequence
-        - "chunk_gated_delta_rule": GDN chunked scan (core recurrence)
-      Generation phase:
-        - "causal_conv1d_update": Single-step causal conv state update
-        - "fused_sigmoid_gating_delta_rule_update": Single-step GDN recurrence
-
-    Uses full (unsharded) dimensions for database lookup; collector data is per-layer.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        kernel_source: str,
-        phase: str,
-        d_model: int,
-        num_k_heads: int,
-        head_k_dim: int,
-        num_v_heads: int,
-        head_v_dim: int,
-        d_conv: int,
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._kernel_source = kernel_source
-        self._phase = phase
-        self._d_model = d_model
-        self._num_k_heads = num_k_heads
-        self._head_k_dim = head_k_dim
-        self._num_v_heads = num_v_heads
-        self._head_v_dim = head_v_dim
-        self._d_conv = d_conv
-        self._weights = 0.0
-
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        batch_size = kwargs.get("batch_size")
-        s = kwargs.get("s")
-        seq_len = s if self._phase == "context" else None
-        result = database.query_gdn(
-            phase=self._phase,
-            kernel_source=self._kernel_source,
-            batch_size=batch_size,
-            seq_len=seq_len,
-            d_model=self._d_model,
-            num_k_heads=self._num_k_heads,
-            head_k_dim=self._head_k_dim,
-            num_v_heads=self._num_v_heads,
-            head_v_dim=self._head_v_dim,
-            d_conv=self._d_conv,
-        )
-        return PerformanceResult(
-            latency=float(result) * self._scale_factor,
-            energy=result.energy * self._scale_factor,
-            source=getattr(result, "source", "silicon"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-
-class Mamba2(Operation):
-    """
-    Mamba2 operation for NemotronH hybrid models.
-
-    Models the Mamba2Mixer layer which consists of:
-    - in_proj: Linear projection (hidden_size -> expanded_size)
-    - conv1d: Causal 1D convolution
-    - SSM: Selective State Space Model (scan operation)
-    - norm: RMSNorm with gating
-    - out_proj: Linear projection back to hidden_size
-
-    This is a SOL-based approximation that models:
-    - Two GEMMs for in_proj and out_proj
-    - Memory operations for conv1d and SSM scan
-
-    The internal state dimension is calculated as:
-    expanded_size = 2 * (nheads * head_dim + 2 * n_groups * d_state)
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        hidden_size: int,
-        nheads: int,
-        head_dim: int,
-        d_state: int,
-        d_conv: int,
-        n_groups: int,
-        chunk_size: int,
-        tp_size: int,
-        quant_mode: common.GEMMQuantMode,
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._hidden_size = hidden_size
-        self._nheads = nheads
-        self._head_dim = head_dim
-        self._d_state = d_state
-        self._d_conv = d_conv
-        self._n_groups = n_groups
-        self._chunk_size = chunk_size
-        self._tp_size = tp_size
-        self._quant_mode = quant_mode
-
-        # Calculate dimensions matching TensorRT-LLM mamba2_mixer.py lines 76-78:
-        # d_inner = head_dim * nheads
-        # d_in_proj = 2 * d_inner + 2 * n_groups * d_state + nheads
-        # conv_dim = d_inner + 2 * n_groups * d_state
-        self._d_inner = nheads * head_dim
-        self._conv_dim = self._d_inner + 2 * n_groups * d_state
-        self._in_proj_out_size = 2 * self._d_inner + 2 * n_groups * d_state + nheads
-
-        # Calculate weights (in_proj + conv1d + out_proj + A + D + dt_bias + norm)
-        # in_proj: hidden_size * in_proj_out_size (Linear d_model -> d_in_proj)
-        # conv1d: d_conv * conv_dim (Linear d_conv -> conv_dim, stored as Linear for TP)
-        # out_proj: d_inner * hidden_size (Linear d_inner -> d_model)
-        # A, D, dt_bias: nheads each (small, ignored for weight calculation)
-        # norm: d_inner (small, ignored)
-        self._weights = (
-            (
-                hidden_size * self._in_proj_out_size  # in_proj
-                + d_conv * self._conv_dim  # conv1d
-                + self._d_inner * hidden_size  # out_proj
-            )
-            * quant_mode.value.memory
-            // tp_size
-        )
-
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        """
-        Query Mamba2 latency using SOL-based approximation.
-
-        Models the operation as:
-        1. in_proj GEMM: (x, hidden_size) @ (hidden_size, in_proj_out_size)
-        2. conv1d: Memory-bound operation
-        3. SSM scan: Memory-bound recurrent operation
-        4. out_proj GEMM: (x, d_inner) @ (d_inner, hidden_size)
-        """
-        x = kwargs.get("x")  # num tokens
-
-        # Apply TP sharding (matching TensorRT-LLM mamba2_mixer.py lines 81-84)
-        # tp_nheads = nheads // tp_size
-        # tp_d_inner = d_inner // tp_size
-        # tp_ngroups = n_groups // tp_size
-        # tp_conv_dim = conv_dim // tp_size
-        nheads_per_gpu = self._nheads // self._tp_size
-        d_inner_per_gpu = nheads_per_gpu * self._head_dim
-        n_groups_per_gpu = self._n_groups // self._tp_size
-        conv_dim_per_gpu = d_inner_per_gpu + 2 * n_groups_per_gpu * self._d_state
-        in_proj_out_per_gpu = 2 * d_inner_per_gpu + 2 * n_groups_per_gpu * self._d_state + nheads_per_gpu
-
-        total_latency = 0.0
-        total_energy = 0.0
-
-        # 1. in_proj GEMM: hidden_size -> in_proj_out_size
-        in_proj_result = database.query_gemm(x, in_proj_out_per_gpu, self._hidden_size, self._quant_mode)
-        total_latency += float(in_proj_result)
-        total_energy += in_proj_result.energy
-
-        # 2. conv1d: Memory-bound operation on conv_dim (not just d_inner)
-        # conv1d operates on xbc which has dimension conv_dim
-        # Read: x * conv_dim * d_conv (for conv states) + x * conv_dim (input)
-        # Write: x * conv_dim (output)
-        conv_read_bytes = x * conv_dim_per_gpu * (self._d_conv + 1) * 2  # bfloat16
-        conv_write_bytes = x * conv_dim_per_gpu * 2
-        conv_result = database.query_mem_op(conv_read_bytes + conv_write_bytes)
-        total_latency += float(conv_result)
-        total_energy += conv_result.energy
-
-        # 3. SSM scan: Memory-bound recurrent operation
-        # For prefill (context), uses chunked scan
-        # For decode (generation), uses selective_state_update
-        # Approximate as memory operation:
-        # Read: x * (d_inner + n_groups * d_state * 2 + nheads) for x, B, C, dt
-        # Write: x * d_inner for output
-        ssm_read_bytes = (
-            x
-            * (
-                d_inner_per_gpu
-                + n_groups_per_gpu * self._d_state * 2  # B and C
-                + nheads_per_gpu  # dt
-            )
-            * 2
-        )
-        ssm_write_bytes = x * d_inner_per_gpu * 2
-        ssm_result = database.query_mem_op(ssm_read_bytes + ssm_write_bytes)
-        total_latency += float(ssm_result)
-        total_energy += ssm_result.energy
-
-        # 4. norm: RMSNormGated on d_inner (TRT-LLM mamba2_mixer.py line 315)
-        # Read SSM output, apply norm with gating, write normalized output
-        norm_read_bytes = x * d_inner_per_gpu * 2  # bfloat16
-        norm_write_bytes = x * d_inner_per_gpu * 2  # bfloat16
-        norm_result = database.query_mem_op(norm_read_bytes + norm_write_bytes)
-        total_latency += float(norm_result)
-        total_energy += norm_result.energy
-
-        # 5. out_proj GEMM: d_inner -> hidden_size
-        out_proj_result = database.query_gemm(x, self._hidden_size, d_inner_per_gpu, self._quant_mode)
-        total_latency += float(out_proj_result)
-        total_energy += out_proj_result.energy
-
-        # Merge sources from every sub-result so the composite reflects mixed
-        # silicon/empirical provenance instead of defaulting to silicon.
-        sub_sources = [
-            getattr(r, "source", "silicon")
-            for r in (in_proj_result, conv_result, ssm_result, norm_result, out_proj_result)
-        ]
-        merged_source = sub_sources[0] if all(s == sub_sources[0] for s in sub_sources) else "mixed"
-
-        return PerformanceResult(
-            latency=total_latency * self._scale_factor,
-            energy=total_energy * self._scale_factor,
-            source=merged_source,
-        )
-
-    def get_weights(self, **kwargs):  # Mamba2 weights
-        return self._weights * self._scale_factor
-
-
 class DeepSeekV4MHCModule(Operation):
     """DeepSeek-V4 manifold-constrained hyper-connection pre/post module."""
